chore: remove redundant delegate, move service calls
