+++ src/xcode/ENA/ENA/Source/Protocols/HomeCardCellButtonDelegate.swift
@@ -1,6 +1,7 @@
+//
 // Corona-Warn-App
 //
-// SAP SE and all other contributors
+// SAP SE and all other contributors /
 // copyright owners license this file to you under the Apache
 // License, Version 2.0 (the "License"); you may not use this
 // file except in compliance with the License.
@@ -14,18 +15,10 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+//
 
-import UIKit
+import Foundation
 
-class SubmitCollectionViewCell: HomeCardCollectionViewCell {
-	@IBOutlet var iconImageView: UIImageView!
-	@IBOutlet var titleLabel: UILabel!
-	@IBOutlet var bodyLabel: UILabel!
-	@IBOutlet var contactButton: UIButton!
-
-	weak var delegate: HomeCardCellButtonDelegate?
-
-	@IBAction func submitButtonTapped(_: UIButton) {
-		delegate?.buttonTapped(cell: self)
-	}
+protocol HomeCardCellButtonDelegate: class {
+	func buttonTapped(cell: HomeCardCollectionViewCell)
 }
